LayoutTests/crypto/subtle/gc.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>

description("Test that window.crypto.subtle wrapper preserves custom properties.");
jsTestIsAsync = true;

function startTest()
{
    window.crypto.subtle.foo = "bar";
    gc();
    setTimeout(continueTest, 10);
}

function continueTest()
{
    gc();
    setTimeout(finishTest, 10);
}

function finishTest()
{
    gc();
    shouldBe('window.crypto.subtle.foo', '"bar"');
    finishJSTest();
}

window.onload = startTest;

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
